+++ mail/mail-uniffi/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2025-01-XX
 
+### Fixed
+  - [ET-1863] Mailbox counter watchers are updated whenever user marks conversation/message as read/unread
+  - [ET-1863] System labels (like Outbox) watchers are updated whenever labels change
+
 ## [0.37.2] - 2025-01-27
 
 ### Fixed